FakeSeries class created as a test helper for chart series.
DefaultNavigation...ActionListener class and DefaultShiftNavi...ChangeListener class created which will serve as outer listeners to various types of charts.
This is done to enable different charts behaviours for shift values and moving left or right, i.e. CandleStick charts that use DateRange in ther DomainAxis will behave differently to standard XYPlots, because their units will be calculated in a different way.
The classes have extracted to avoid awkward switch...case statements for Charts to choose the proper listeners. Now this can be done by factories, for instance, or in any other way by the developers.